fix non-fixed full width bootstrap navbar
Content in the navbar should not be pushed out to the edges of the
browser, it should be lined up to the grid like the rest of the site.

This ultimately should be fixed in bootstrap, but they don't seem to
think it's a problem - https://github.com/twitter/bootstrap/issues/3476.

The hooligan theme would be another good place to fix it, I'll look
into that as well.

I also added a Gemfile to manage dependencies. Install bundler (should
already have it) and then do a bundle install. Everything else should
work as normal.